AstronomyGeometry

The GenerateSatellites namespace within the QuantumSatelliteTools.jl provides tools for generating vectors of TLE or satellite propagator objects.

Getting Started

Add the QuantumSatelliteTools.jl package using Julia's package manager:

] add QuantumSatelliteTools

Import the GenerateSatellites namespace into your Julia environment:

using QuantumSatelliteTools.GenerateSatellites

Index

QuantumSatelliteTools.GenerateSatellites.generate_regular_array

QuantumSatelliteTools.GenerateSatellites.generate_regular_array_TLEs

QuantumSatelliteTools.GenerateSatellites.get_active_satellite_TLEs

QuantumSatelliteTools.GenerateSatellites.get_active_satellites

QuantumSatelliteTools.GenerateSatellites.refresh_satcat

QuantumSatelliteTools.GenerateSatellites.save_TLEs

QuantumSatelliteTools.GenerateSatellites.update_starlink_TLEs

API Documentation

QuantumSatelliteTools.GenerateSatellites.generate_regular_array — Method

Generate synthetic propagators for a regular constellation of satellites.

Keyword Arguments

name_prefix::String="SAT": Prefix for satellite names.

orbital_planes::Int=12: Number of orbital planes.

sats_per_plane::Int=18: Satellites per orbital plane.

altitude_km::Int=500: Altitude of the orbits in kilometers.

inclination_rad::Float64=π/2: Inclination angle in radians.

frozen_orbits::Bool=false: Whether to use frozen orbit parameters. See https://juliaspace.github.io/SatelliteAnalysis.jl/stable/man/frozen_orbits/.

Returns

Vector{Propagator}: Vector of synthetic Propagator objects.

QuantumSatelliteTools.GenerateSatellites.generate_regular_array_TLEs — Method

Generate synthetic TLEs for a regular constellation of satellites.

Keyword Arguments

name_prefix::String="SAT": Prefix for satellite names.

orbital_planes::Int=12: Number of orbital planes.

sats_per_plane::Int=18: Satellites per orbital plane.

altitude_km::Int=500: Altitude of the orbits in kilometers.

inclination_rad::Float64=π/2: Inclination angle in radians.

frozen_orbits::Bool=false: Whether to use frozen orbit parameters. See https://juliaspace.github.io/SatelliteAnalysis.jl/stable/man/frozen_orbits/.

Returns

Vector{TLE}: Vector of synthetic TLEs.

QuantumSatelliteTools.GenerateSatellites.get_active_satellite_TLEs — Method

Retrieve TLEs of active payload satellites from satcat.csv and Celestrak.

Keyword Arguments

OBJECT_NAME::String="": Optional filter to match part of the satellite name.

row_filter::Function=sat_row->true: Function to filter rows from satcat.csv.

return_partial::Bool=true: If true, returns partial results even if some fetches fail.

verbose=false: If true, prints progress info.

Returns

Vector{TLE}: A vector of fetched TLE objects.

QuantumSatelliteTools.GenerateSatellites.get_active_satellites — Method

Retrieve and return Propagator objects of active payload satellites from satcat.csv and Celestrak.

Keyword Arguments

OBJECT_NAME::String="": Optional filter to match part of the satellite name.

row_filter::Function=sat_row->true: Function to filter rows from satcat.csv.

return_partial::Bool=true: If true, returns partial results even if some fetches fail.

verbose=false: If true, prints progress info.

Returns

Vector{Propagator}: A vector of fetched Propagator objects.

QuantumSatelliteTools.GenerateSatellites.refresh_satcat — Method

Download the latest satcat.csv catalog from Celestrak and save it locally.

QuantumSatelliteTools.GenerateSatellites.save_TLEs — Method

Save a vector of TLE objects to a specified file.

Arguments

TLE_vector::Vector: The vector of TLE objects to save (see SatelliteToolboxTle.jl).

file_path::String: The path to the output file.

QuantumSatelliteTools.GenerateSatellites.update_starlink_TLEs — Method

Update the stored Starlink TLEs by querying Celestrak for new entries. Minimizes calls to Celestrak by only querying for Starlink satellites that are listed in satcat.csv as active but do not appear in the local Starlink TLE database.

TODO: remove inactive satellites

Keyword Arguments

verbose::Bool=false: If true, prints status messages during fetching.